+++ CHANGELOG.md
@@ -14,6 +14,9 @@ settings `search_results_page_filtering_enabled` and `search_results_page_filter
 - Use `flat` buttons for search and for confirming quitting the app.
 - Refactor the logic to navigate search result pages to use dynamic actions.
 - Add support for client-side certificates and SSL configuration
+- Add support for Bearer authentication
+- Add support for configuring bearer authentication via config variable `use_bearer_authentication`.
+- Add configuration variable `cloud` to distinguish between on-premises and cloud platform.
 
 ### Bug Fixes